Fix incomplete config read and add lock on state
Connector configurations were being read to identify the presence of
the connectors and were assumed to have valid configuration data, but
they were sometimes incomplete since the connector had not yet been
enabled.  This has been fixed for now by rereading the configuration
after confirming the connector is enabled.  It could be improved by
not reading the entire configuration the first time, since the purpose
is only to identify connectors.

Also, a mutex has been added to the server state to avoid potential
data races.